+++ src/main/java/com/gabo/videoClub/repositories/IMovieRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IMovieRepository extends JpaRepository<Movie, Long> {
+public interface IMovieRepository extends JpaRepository<Movie, Integer> {
 }
