+++ src/main/java/com/gabo/videoClub/services/IClientService.java
@@ -5,9 +5,9 @@ import com.gabo.videoClub.dto.responses.ClientResponseDto;
 
 import java.util.List;
 
-public interface IClientService extends IHateoas <Long> {
+public interface IClientService extends IHateoas <Integer> {
     public ClientResponseDto save(ClientRequestDto clientResponse);
     public List<ClientResponseDto> getAllClients();
-    public ClientResponseDto getClientById(Long id);
-    public void deleteClient(Long id);
+    public ClientResponseDto getClientById(Integer id);
+    public void deleteClient(Integer id);
 }
